Working with json completed

+++ remember_me.py
@@ -1,10 +1,13 @@
 from pathlib import Path
 import json
 
-username = input("What is your name? ")
-
 path = Path("username.json")
-contents = json.dumps(username)
-path.write_text(contents)
-
-print(f"we'll remember you when you come back, {username}")
+if path.exists():
+    contents = path.read_text()
+    username = json.loads(contents)
+    print(username)
+else:
+    username = input("what is your username? ")
+    contents = json.dumps(username)
+    path.write_text(contents)
+    print(f"we'll remember you when you come back, {username}")
